Allow the many side of a one-to-many relatonship to be a non-KeyedEntity.
We accomplish this by requiring implicit evidence of KeyedEntity-ship only if
the left.associate(right) method of the OneToManyRelationship is called. In
that case, right is automatically inserted or updated, which requires that it
be a KeyedEntitiy.

Now it is possible to obtain the foreign key benefits for non-KeyedEntities,
assuming the user does not use associate, but manages adding rows to the table
on their own (a quite common use case, I imagine).