Closes #93: Add secondary constructors in custom types
This commit replaces the algorithm of _createCustomTypeFactory in FieldMetaData.
The former version takes the first one-arg constructor in the field class
for which it can create a default value and returns a closure for it.
This algorithm fails if the field's class has multiple one-arg constructors
for different types.

The new version of the algorithm is able to handle classes with multiple
one-arg constructors. It chooses the best constructor for the given
type to create an instance.