Create SquerylSQLException and throw for DB errors
This is a proposed fix for #83.

By throwing a specific exception type to indicate database errors, we
can make it easier for client code interested in detecting/handling
database errors to distinguish those exceptions from other exceptions
that may be thrown (either by Squeryl or other code in the same
try-catch block).  This commit introduces the SquerylSQLException type
for this purpose.

Incompatibilities are minimized by extending SquerylSQLException from
RuntimeException and keeping the messages mostly unchanged.

Possible incompatibility due to changing the message in Table.{insert,
update}.  Since this error could result from poorly written triggers
which increase the row count unnecessarily, users may be confused if the
operation was successful yet the exception was thrown without a clear
reason why.  If this is controversial, the commit can be reworked to
avoid this change.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>