Proactively remove any ThreadLocal state
Proactively remove any ThreadLocal state when cleaning up a Session. For
reasons still exceedingly mysterious, the ThreadLocal state used when
building the AST is very rarely not cleaned up. This can corrupt a
subsequent query that uses the same thread (potentially much later in
the case of a thread pool). Work around this by removing any lingering
ThreadLocal state at the end of a transaction.