A few changes to get travis publishing.
My sonatype credentials are now encrypted and stored in the .travis.yml file.  Outside of Travis, credentials can be passed in using the environment variable SONATYPE_USERNAME & SONATYPE_PASSWORD.  I've also removed the sbt stuff that existed inside the repo, since we can't rely on it if we need to ensure builds work properly from a CI environment.